legacy.py: Handle missing weight or flow data
Previously the algorithm to emulate the de1app behavior would try to get
the previous sample value, even if there was not value at that index.
This would raise an IndexError that apparently would get swallowed
by the executor. The shot would not be uploaded to Visualizer.

Reported and identified by Moo.