RESOURCE v3.8.0, MAPPING v4.2.0 for Firmware Versions 1315-1352
Versions 1315-1352 include new ways of managing the fill kit with
a user-present flag. It also includes the ability to shut off
the USB power outlet, used to charge the tablet's battery.
Although marketed as a power-saving feature, the reality it that
the Teclast tablets' batteries swell to the point of deforming
the tablet case. Version 1352 wisely re-enables the USB power
10 minutes after disabling it.

These are all only "activated" for Firmware 1320 and later.

Increased config.de1.MAX_WAIT_FOR_READY_EVENTS = 4.0
(was 3.5 seconds for 25 events, now 28 events needed)

    MMR
         STEAM_PURGE_MODE        = 0x3850
         ALLOW_USB_CHARGING      = 0x3854
         APP_FEATURE_FLAGS       = 0x3858
         REFILL_KIT_PRESENT      = 0x385c
         USER_PRESENT            = 0x3860

    FeatureFlags
         steam_purge_mode
         allow_usb_charging
         app_feature_flag_user_present
         refill_kit_present
         user_present

The USER_PRESENT MMR can be written to, typically with True,
to indicate that there is a user interacting with the machine.
If the APP_FEATURE_FLAGS & AppFeatureFlag.USER_PRESENT, the DE1
expects the app to be able to process the Idle, UserNotPresent
state and substate, as well as enabling the fill kit only when
the user is present. The USER_PRESENT MMR is cleared by the DE1
two minutes after it is set. Client applications should not read
the USER_PRESENT MMR at a high rate as it is read on every query.
The ALLOW_USB_CHARGING MMR also is read on every query as it also
resets on its own.

Closes #30