de1: Provide config.de1.PATCH_ON_CONNECT
When a DE1 connects, the contents of config.de1.PATCH_ON_CONNECT
are submitted as a PATCH to the DE1 through the inbound HTTP API.

At this time it is done immediately after _notify_ready() as "ready"
is a prerequisite for processing.

The requests library is synchronous, so it can't be directly called in line
in the thread that would need to execute the patching.