MAPPING 4.2.1: Include last-updated time for DE1 state
It is possible that the response to a request to the API for the DE1 state
and an MQTT report of the state arrive "out of sequence". Previously
there was no way to determine which was the most recent of the two.

This patch adds the last-updated time of the state to the API response.

    {
        "state": {
            "last_updated": 1661999107.5938282,
            "state": "Sleep",
            "substate": "NoState"
        }
    }

Combined with the times present in the MQTT notifications, the ambiguity
can be resolved.

Closes #32