legacy_to_json: Provide command-line utility to convert profiles
Parses the undocumented Tcl format of de1app as of v1.40.1 (20220-09-04)

Command-line help is available with the -h option

legacy_to_json.py -h
usage: legacy_to_json.py [-h] [-a AUTHOR] [-i INPUT | -v VISUALIZER] \
                         [-o OUTPUT] [-d DIR] [-f]

Executable to open a Tcl profile file and write as JSON v2.1.
Input and output default to STDIN and STDOUT

options:
  -h, --help                               show this help message and exit
  -a AUTHOR, --author AUTHOR               Replace author
  -i INPUT, --input INPUT                  Input file
  -v VISUALIZER, --visualizer VISUALIZER   Visualizer URL
  -o OUTPUT, --output OUTPUT               Output file
  -d DIR, --dir DIR                        Output directory
  -f, --force                              Overwrite if output exists

Input can be stdin, a file specified with -i, or a Visualizer URL,
as copied from the visualizer.coffee website, specified with -v
At this time, only the download URL is supported, not the shot URL
or the short code.

Output can be to stdout, or specified with either or both
-o for a file name and -d for a directory.

A file name with no directory is relative to the working directory.
A directory with no file name will take the name from the input,
if available, the profile title (sanitized) as a fall back,
with an extension of .json instead of any it previously had.
Both directory and file name will be concatenated.

Files are not overwritten unless -f is given.

The author field of the profile is replaced by providing -a

Writes a JSON format version "2.1" that is believed to be compatible with
version "2" (presently written, but not read by de1app)
with the following changes:

* Values are typed and represented in JSON as such

* A creator section is added:

    "creator": {
        "name": "legacy_to_json.py",
        "version": "0.0.1",
        "timestamp": "2022-09-09T13:29:27-07:00"
    }

* Cruft based on other parameter values have been dropped,
  such as "flow" when a "pressure" profile and zero-valued
  parameters that indicate "not active"