Implement move-on by weight
Add a feature that will call skip_to_next (frame)
when the weight change in a frame exceeds the "weight" limit,
if one is supplied for that frame.

Parsing is done from JSON v2 and v2.1 profile definitions from the
"weight" element in each frame, if it is present.

The list of values is interpreted as starting with "frame 0"
and incrementing from there. The length of the list may be shorter
than the number of frames. The list may be the empty list.
Values may be positive numbers or None if no limit.

The StopAtNotification has been updated to version 1.1.0 as it adds
the current frame, which may be None/null, as well as the action of
'move on by weight'.

The mode-control objects within the FlowSequencer have been refactored.
Any code accessing these directly should examine both flow_sequencer.py
and flow_sequencer_impl.py.

Frame changes are now logged at the INFO level, including the previous
and current frames, as well as an estimate of the weight.

An estimate of the scale weight is now available as the ScaleProcessor's
current_weight property. This may be None if the last update was over
one second prior.

The Skale II implementation now returns None for current_weight() rather
than raising NotImplementedError. There is not a clear way to request
a weight update from the Skale II on demand, other than notifications.

The signature of the stop_at_notify() method of the FlowSequencer has been
changed to capture information at the time of the call, rather than when
the call is serviced. With skip_to_next() being called, it is possible
that the time of service could be in a different frame of the profile.

The FlowSequencer internal stop_at_weight routine has been renamed to
act_on_weight and handles both SAW and move-on by weight.