Scale: current_weight can return None and does so for Skale II
The current_weight property was intended to be a read-now command
that would return the weight even if the scale was not streaming.

At least in the case of the Skale II and Acaia Lunar 2021, there
does not appear to be a command to support that operation.

Rather than raise NotImplementedError, Skale II now returns None