SAW/MOW: Adjust mass-flow estimate when too high or negative
A bump of the cup or scale can cause a brief spike in weight, either
up or down. As that weight spike moves past the estimator, there will be
a positive and a negative spike in the mass-flow estimate. The positive
spike can trigger SAW or MOW early. The negative spike could delay action.

This commit introduces two, new config parameters:

    config.de1.ACT_AT_WEIGHT_FLOW_THRESHOLD  -- default 10.0 (g/s)
    config.de1.ACT_AT_WEIGHT_FLOW_MULTIPLIER -- default 1.1

If the mass-flow estimate is over ACT_AT_WEIGHT_FLOW_THRESHOLD
then the reported DE1 group flow * ACT_AT_WEIGHT_FLOW_MULTIPLIER is used.

If the mass-flow estimate is negative, 0 is used.

The choice of 0 for negative flows was selected as early drops can cause
negative mass-flow estimates as they hit the cup. The DE1 group flow may
be much higher than the rate of dripping in early frames. Use of the
DE1 group flow could cause early termination of these frames.