SAW/MOW: Adjust mass-flow estimate when too high or negative
A bump of the cup or scale can cause a brief spike in weight, either
up or down. As that weight spike moves past the estimator, there will be
a positive and a negative spike in the mass-flow estimate. The positive
spike can trigger SAW or MOW early. The negative spike could delay action.

This commit introduces new config parameters:

        config.de1.bump_resist.FLOW_THRESHOLD = 10.0  # g/s
        config.de1.bump_resist.FLOW_MULTIPLIER = 1.1
        config.de1.bump_resist.SUB_MEDIAN_WEIGHT = True  # when excessive flow
        config.de1.bump_resist.USE_MEDIAN_WEIGHT_ALWAYS = False
        config.de1.bump_resist.USE_MEDIAN_FLOW_ALWAYS = False

If the mass-flow estimate is over FLOW_THRESHOLD then the reported
DE1 group flow * FLOW_MULTIPLER is used for SAW/MOW. If SUB_MEDIAN_WEIGHT
is true, then the median-weight estimate is used as well.

If the mass-flow estimate is negative, 0 is used.

The choice of 0 for negative flows was selected as early drops can cause
negative mass-flow estimates as they hit the cup. The DE1 group flow may
be much higher than the rate of dripping in early frames. Use of the
DE1 group flow could cause early termination of these frames.