Managed bleak client/device introduced and utilized
In order to provide the ability to easily disconnect from a device
and automatically reconnect when it becomes available again,
the management of Bluetooth devices was restructured.

This enables smoother use of non-dedicated devices, such as
Acaia scales and BlueDOT thermometers.

Rather than managing connection at the Bluetooth level with connect
and disconnect, devices are managed with capture, release, and forget.
A device that has been released is still known by pyDE1. A subsequent
request to capture will wait for the device to be seen if it is not
already sending out advertisement packets. After connection the device
will be initialized and then announced as "ready" to the rest of the
pyDE1 system.

The retry interval for a not-seen device is three-tiered:

  self._retry_start_initial_delay = 300 # seconds
  self._retry_initial_delay = 15 # seconds
  self._retry_start_long_delay = 1800  # seconds
  self._retry_long_delay = 60 # seconds

The delay is in addition to the default scan time of 5 seconds
(override still available through config.bluetooth.SCAN_TIME)
The hold-off is not confirable at this time.

The API endpoint has changed to ``de1/availability``.

  The ``de1/connectivity`` endpoint and ``ConnectivityChange``
  notifications have been deprecated. They will be emulated through
  June, 2023 for backward compatibility with existing clients.

In order to simplify consumers of scale or thermometer data, the
objects that represent these remain constant as viewed by those
consumers. For example, events that have been subscribed to do not
need to be unsubscribed and resubscribed when the underlying device
changes.

See "What's New in Version 2" for additional discussion.

MAPPING 5.1.0 -- source_data available through API

  NB: This is not as robust as desired and may change in the future

  make-source-data.sh can be used as a git hook and the variable described
  can be obtained through the version API

    "source_data": {
        "git": {
            "author_date": "Sun Oct 16 20:32:41 2022 -0700",
            "commit_date": "Sun Oct 16 20:32:41 2022 -0700",
            "hash": "099fbb6",
            "ref_names": "HEAD -> devel-2022-10",
            "subject": "MAPPING 5.1.0 -- source_data available through API"
        }
    }

  `git status --porcelain` would indicate if there are uncommitted changes
  or untracked files that aren't in .gitignore. However, when used as a
  git hook, it will fail to show changes after the commit or checkout.
  Rather than provide a potentially misleading "clean" indication,
  the functionality is not enabled.

MAPPING 5.2.0 -- Allow None for steam temperature

  Steam temperature can be set to 0 to indicate that the steam heater
  should not be powered up to maintain a value in its 140-160 C range.
  Although the setter permitted None, encoding to a u8 did not convert
  None to 0.

MAPPING 5.4.0 -- Add availability
RESOURCE 3.9.0

  DE1_AVAILABILITY = 'de1/availability'
  SCALE_AVAILABILITY = 'scale/availability'

  The returns include both the "mode", as well as "mqtt" which is the
  last-sent packet at the time the request was made.

  The mode can return the value of
  class DeviceAvailabilityState (enum.Enum):

    INITIAL = 'initial'
    UNKNOWN = 'unknown'
    CAPTURING = 'capturing'
    CAPTURED = 'captured'
    READY = 'ready'          # "Ready for use"
    NOT_READY = 'not ready'  # Was READY, but is no longer
    RELEASING = 'releasing'
    RELEASED = 'released'

  The parameters for requesting a change are

    CAPTURE = 'C'
    RELEASE = 'R'

SubscribedEvent: Refactor for bound methods and to use weakrefs

  Previously the subscriber list held direct references to the callbacks.
  These callbacks often contained references to other objects. This caused
  circular reference loops to occur, which potentially could cause problems
  with GC. By using weakrefs, these references to the coroutine functions
  are no longer considered by GC and should allow simplification
  of the process of, for example, replacing one scale with another.

  Should the referent of a weakref be discovered to have vanished
  in the process of sending a notification, it is logged and
  scheduled for removal from the subscriber list.

  Prior to this refactor, only unbound coroutine functions could be used
  as subscribers. This refactor now permits regular functions and coroutine
  functions, bound or not. This should allow simplification of callbacks
  as bound methods can be used in place of closures.

  Closes #25