Visualizer: Provide timeout on POST with requeue
The service to upload to Visualizer was observed to stop uploading
without a reason evident in the logs. It was hypothesized that
the connection was "stalled", leaving the upload unprocessed
and subsequent uploads not retrieved from the pending-upload queue.

A timeout on the request was added with re-queueing if the timeout triggers.

Cosmetic changes were also made to the log messages on queueing
of a shot to be uploade and on retrieval from the queue.