log/record MQTT "message" is now unformatted
Prior to this change, the log/record MQTT payloads had the message
already formatted. As a result consumers that wanted to render it
in a different way would need to parse out the "raw" message itself.

By deferring formatting to the time of sending the MQTT packets
the log/record packet now contains the unformatted message,
along with the other fields.

log -- formatted string:

  2022-11-19 15:47:17,862 INFO [MainProcess] root:
      Started <logging.handlers.QueueListener object at 0x75b6ed90>

log/record -- JSON with unformatted message

  {"version": "1.0.0", "created": 1668901637.862185,
  "levelname": "INFO", "levelno": 20,
  "message": "Started <logging.handlers.QueueListener object at
                 0x75b6ed90>",
  "name": "root", "process": 29531, "processName": "MainProcess",
  "thread": 1996083264, "threadName": "MainThread"}