Read-back on MQTT {TOPIC_ROOT}/update/{subtopic} after change
To enable multiple, concurrent clients without putting excessive load
on the HTTP server, broadcast changes made. This should be able to
replace most API queries except at start-up or resync on returning
to the page.

Any time a PATCH or PUT (state-modifying request) is processed,
an MQTT packet is sent out with the content of the impacted area.

At this time those areas include:
    * de1/profile/id
    * de1/control
    * de1/setting
    * de1/calibration

The retain flag was considered for these, but has not been added
as the broker might be down or have been down, so a retained packet
may not be available or may not represent current state.

This implementation is likely to be refactored. The functionality
as seen from outside is expected to be backwards compatible.