Update acaia.py to GenericScale approach, adjust for pre-2021 Lunar
Examining behavior of the pre-2021 Lunar with the Acaia app confirms
that the scale will time out even with all three "heartbeat" packets
being sent. As the heartbeat is not needed by this implementation,
it is not enabled.

Earlier testing was done with a Lunar 2021, which is not available
for testing at this time.