Acaia: Provide set_beep(audible:bool)
As an initial step to trying to silence the beeping from tare requests
at the start of each shot, implement the command to set the beep on or off.

Examination of the behavior of the Acaia Lunar suggested that it would beep
any time the beep was set to enabled. There also didn't seem to be any
ACK packet sent by the scale with turning off the beep, only when turning on.
No method for triggering a status update (efdd 08...) has been identified.
Based on this, it seems that it is not possible to silence the scale
and then restore it silently. As a result, no further development
to silence the scale during hold-at-tare is planned.

Some improvements in logging messages around hold-at-tare were clarified
as they used the word "release". Status as reported by the Acaia scale
is now retained, but not used at this time.