Database Schema 3 for capture of DeviceAvailability and ScaleChange
Upgrade from Schema 2 to 3 is performed automatically. A backup is made
and left in the same directory as the "live" database.

The BACKUP_TIMEOUT is set to 120 seconds, which seems adequate
for a 500 MB / 3000 sequence database on a Raspberry Pi 3B+.
Should it not be adequate, the timeout can be lengthened through config.

Do not try to compress the backup. The backup is only occasionally done
as the schemas should be relatively stable at this point. It is intended
to be a snapshot in case something goes wrong with the update. Post-facto
compression doesn't resolve low-on-space problems and does add to writes.
At this time, it is not a priority to work out how to write the call
so that it doesn't terminate if pyDE1 terminates, such as from a failed
database upgrade.

This commit also includes the database changes needed for the BlueDOT
thermometer and stop-at-temperature, added in later commits.