Change wait-for-shutdown actions to be async def
When testing automated schema upgrades, an exception in check_schema()
was not completing shutdown as expected. The exact mechanism is not
understood, but it was related to running on_shutdown_underway_cleanup()
in the thread pool executor. It is possible that it was hanging
as the asyncio loop had not been started yet.

A multiprocessing.Event.wait() is not an awaitable. mp_event_wait()
is supplied as an awaitable wrapper. It operates by running the
multiprocessing .wait() in the default executor. The loop periodically
checks for either timeout or if the "abandon_on_event" event is set.
It is thought that this resolves the problem by deferring the creation
of the task in the executor until the loop is running.

This has the advantage that the cleanup process can be async def-ed.