Remove convenience imports from event_manager/__init__.py
In an attempt to simplify importing of portions of the pyDE1 system,
the imports in event_manager/__init__.py were removed. These allowed
importing at the event_manager level without needing to specify which
file within the package held the desired functionality. Unfortunately,
this did not resolve the "partially intialized module" import errors.

Retaining the change as it does not change function and it may
simplify future refactor of the module/import structure.

Removed:

    from .event_manager import (
       SubscribedEvent, EventWithNotify, SequencerGate, send_to_outbound_pipes,
    )
    from .payloads import (
        EventPayload, EventNotificationName, SequencerGateName,
        SequencerGateNotification,
    )