Add a delay when trying to connect two devices simultaneously
If two devices are triggered to connect at the same time
a org.bluez.Error.InProgress can be raised. This can happen
when the DE1 waking up and both the scale and thermometer
are known.

Catch the exception and try again shortly.
Use in-progress hold-off of 0.2 sec. 0.1 sec was too short.