Bluetooth scanner and scanning rework
The approach to Bluetooth scanning was reworked to use changes in the Bleak
Bluetooth library as well as to simplify client code.

Scanning is now by "role", one of

- DE1
- Scale
- Thermometer

The scan results are reported over MQTT and consist of a boolean ``scanning``
indicating if a scan is still underway, and ``devices``, an array of
accumulated information about all devices matching the requested role
seen during the scan. The packet contains
- ``address``
- ``name``
- ``RSSI``
for each device, with updates as additional devices are found during the scan.

The results are no longer retained in the ``DiscoveredDevices`` structure and the APIs
to access that structure are not available.

MAPPING 7.0.0, RESOURCE 5.0.0
-----------------------------

- ``Resource.SCAN`` (``scan``)
  - Now takes a string representing the role
  - Can no longer be PATCH-ed (use PUT with the desired role)

- ``Resource.SCAN_DEVICES`` (``scan_devices``) -- has been removed