steam-to-temp: Use "Idle" if steam_purge_mode is active
With more recent firmware, there is a STEAM_PURGE_MODE MMR
that enables "going to puff mode" when tapping the stop icon
on the GHC. Tests indicate that the same behavior occurs
when sending a request for Idle. If this feature is present
and enabled, use this simpler approach for stopping the steam.