ManagedBleakDevice: Reset retry timer on new capture request
If there is a pending capture with a hold-off, the device will not start
to try to connect until the hold-off expires. This may slow response to
the DE1 coming out of sleep.

Call _retry_reset() when making a capture request to speed the process.

Also remove redundant call to _retry_set_timer() It was already being called
by _start_capture_with_lock()