de1: Various improvements in firmware upload
Allow PUT to de1/firmware/cancel

Don't persist FirmwareUpload packets

internal_type was previously added to the mappings to be able to
convert a simple JSON type into a specified enum. As it can be None,
the issubtype check would fail if that branch of the logic was reached.

Added check for None, made type Optional[type(Enum)]
and removed one misuse with str